+++ tests/commandsAsPleb.test.ts
@@ -21,7 +21,7 @@ describe("Command as pleb", function () {
 		await setIsQueueCreator(true);
 		await setIsQueueAdmin(true);
 		await commandResponseInTestChannel(`${QUEUE_COMMAND} restart`, "queue");
-		await commandResponseInTestChannel(`${QUEUE_COMMAND} limit entry-duration 0`, "removed");
+		await commandResponseInTestChannel(`${QUEUE_COMMAND} limit entry-duration-max 0`, "removed");
 		await commandResponseInTestChannel(`${QUEUE_COMMAND} teardown`, "deleted");
 
 		// Remove the Queue Admin role from the tester bot